feat: add stock class and code sql for his creation

+++ src/db/migration/create_table_dish_ingredient.sql
@@ -1,7 +1,7 @@
 CREATE TABLE dish_ingredient(
     id_dish VARCHAR,
     id_ingredient VARCHAR,
-    required_quantity int,
+    required_quantity FLOAT,
     unity unity,
     PRIMARY KEY (id_dish, id_ingredient),
     FOREIGN KEY (id_dish) REFERENCES dish (id),
